+++ my-app/src/components/cardNumberSection/CardNumberSection.tsx
@@ -1,7 +1,7 @@
-import { useId, useRef, useState } from 'react';
-import { getCardNumberError, isInputValidate } from '../../utils/Validation';
+import { useId } from 'react';
 import CommonSection from '../common/commonSection/CommonSection';
 import NumberInput from '../common/numberInput/NumberInput';
+import { useCardNumber } from './useCardNumber';
 
 interface Props {
   value: string[];
@@ -9,34 +9,9 @@ interface Props {
 }
 
 const CardNumberSection = ({ value, setValue }: Props) => {
-  const [errors, setErrors] = useState<boolean[]>([false, false, false, false]);
-  const inputRefs = useRef<(HTMLInputElement | null)[]>([]);
   const cardNumberIds = useId();
+  const { errors, inputRefs, handleOnChange, handleOnBlur, finalErrorMessage } = useCardNumber({value, setValue});
 
-  function handleOnChange(inputValue: string, index: number) {
-    if (!isInputValidate(inputValue, 4)) return;
-
-    const newValue = [...value];
-    newValue[index] = inputValue;
-    setValue(newValue);
-
-    if (inputValue.length === 4 && index < 3) {
-      inputRefs.current[index + 1]?.focus();
-    }
-  }
-
-  const handleOnBlur = (inputValue:string, index: number) => {
-    const isError = getCardNumberError(inputValue) !== '';
-    setErrors((prev) => {
-      const newErrors = [...prev];
-      newErrors[index] = isError;
-      return newErrors;
-    });
-  }
-
-  const errorIndex = errors.findIndex((isError) => isError);
-  const finalErrorMessage = errorIndex !== -1 ? getCardNumberError(value[errorIndex]) : '';
-  
   return (
     <CommonSection
       title="결제할 카드 번호를 입력해주세요"
